feat: Add robust 'sampling' to willowtree.py
Add two robust sampling algorithms to willowtree.py

This commit adds robust versions of two sampling strategies
'kurtosis-matching' and 'first-partial-moment'.

Robust means that if one of the three cases occur:

- the user inputs a value of gamma not in [0, 1];
- the algorithm name is unrecognised;
- optimisation is unsuccessful with the supplied parameters;

the function automatically:

- sets gamma to the closest bound (i.e. to 0 if initially
  lower than zero, to 1 if initially higher than zero);
- switches to 'kurtosis-matching', the fastest and most
  reliable algorithm;
- performs an additional round of optimisations increasing
  gamma first, and only after a full unsuccessful cycle
  0 -> 1 (if initial gamma ~ 1) or 1 -> 0 (if initial
  gamma ~ 0) raises n, the number of space nodes, and again
  gamma. It does so until a solution is found.

With regards to the last point, if no initial solution is
found, for the first 2 seconds gamma is either increased or
decreased by 1e-9, then, for 8 seconds, by 1e-6, and finally
by 1e-2. This method ensures that, if possible, a value of
gamma closer to the one supplied is first used, and only if
no near solution is found gamma is raised, in absolute value,
by a larger step.