feat: Shorten Markov chain if last matrices bad
Automatically remove bad transition matrices from the Markov chain
if it is impossible to replace them with interpolated versions.

This feature addresses an issue raised with commit 02d210a and
related to the fact that the matrix interpolation algorithm breaks
if the chain ends with badly specified transition matrices.

Now such matrices are automatically scrapped, and a shortened chain
returned. This way, the algorithm breaks no more.

Resolves: #2