fix: Refine constraint on probabilities in LP
Refine the constraint on probabilities to rule out p < 0 and
p > 1 while also accounting for occasional NaN output from the
linear programming algorithm.

The NaN output previously broke the algorithm, since the added
constraints '(P[P < 0]).any()' and '(P[P > 1]).any' raised the
TypeError: 'float' object is not subscriptable.

This commit fixes the error introduced in dec5db5 by checking
whether the output of the LP problem is float (NaN) and:

 - if not, by adding the two constraints on the LP problem;
 - otherwise, by automatically assigning flag -1 (failure) and
   moving on to the next problem.

However, the side effect is an increase in running time.
For this reason, the algorithm is now timed so that it stops
whenever any iteration lasts longer than one minute.

Resolves: #4