feat: Plot tree with no Markov chain if k > 1
Plot willow tree even if no transition matrix in the Markov chain
is well-defined.

This feature addresses an issue which causes the algorithm to break
when linear programming is unable to generate any valid transition
matrix in the chain.

The commit resolves the case k > 1, but leaves unsolved case k = 1.

Resolves: #9